+++ th2_data_services_lwdp/events_tree/__init__.py
@@ -12,5 +12,5 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .events_tree_collection import EventsTreeCollectionProvider
+from .events_tree_collection import EventsTreeCollection
 from .parent_events_tree_collection import ParentEventsTreeCollection