+++ th2_data_services_lwdp/utils/__init__.py
@@ -12,5 +12,5 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .misc import DatetimeConverter, _check_list_or_tuple, _check_milliseconds
-from .pages import Page
+from .misc import _check_list_or_tuple, _check_milliseconds
+from .page import Page